# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "La Mesa API",
    "description": "REST JSON contract for **La Mesa**. Every mutating response carries a full A2UI\nmessage array. No WebSockets or SSE (INV-010).\n\nFrozen routes live in `SPECS.md` §8 (`REQ-API-01` … `REQ-API-08`). The frontend\nconsumes this HTTP API only; LLM providers are not called from the client.\n\nInteractive docs: `/swagger` or `/docs` (Swagger UI) and `/redoc`. Machine-readable spec:\n`/openapi.json`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/message": {
      "post": {
        "tags": [
          "message"
        ],
        "summary": "Send user utterance to agent (REQ-API-02)",
        "description": "Send user text or base64 audio to the AI agent orchestrator (REQ-API-02).\n\n- If `audio_b64` is provided without `text`, audio is transcribed first via the `voice` MCP server.\n- The agent processes the utterance, queries financial tools via MCP, and generates a new A2UI interface.\n- If accessibility is enabled for this persona, `amitie.voz_color.v1` is generated with pre-warmed TTS audio (REQ-ACC-01).\n- Every response returns the complete A2UI message array over HTTP (INV-010).",
        "operationId": "post_message",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent turn executed successfully; returns complete A2UI message array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found in the database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/saving-bags": {
      "post": {
        "tags": [
          "saving-bags"
        ],
        "summary": "Create a saving bag (REQ-API-05)",
        "description": "Create a new saving bag goal (REQ-API-05).\n\n- Persists the saving bag entity in SQLite via the `savings` MCP server.\n- Agent analyzes the goal name (e.g. 'Viaje a Japón') and infers clarifying questions.\n- Emits an interactive questionnaire surface rendered via A2UI.",
        "operationId": "create_saving_bag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingBagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saving bag initialized; returns clarification questions in A2UI form.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingBagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "saving-bags"
        ],
        "summary": "List saving bags (REQ-API-05)",
        "description": "List all saving bags created by a given user.",
        "operationId": "list_saving_bags",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "User identifier whose saving bags to list.",
              "default": "u_ana",
              "title": "User Id"
            },
            "description": "User identifier whose saving bags to list."
          }
        ],
        "responses": {
          "200": {
            "description": "Array of saving bag records.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingBagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/saving-bags/{bag_id}": {
      "get": {
        "tags": [
          "saving-bags"
        ],
        "summary": "Get saving bag details (REQ-API-05)",
        "description": "Retrieve details, current plan, and hydrated UI for a saving bag.",
        "operationId": "get_saving_bag",
        "parameters": [
          {
            "name": "bag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saving bag details and active hydrated surface.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingBagResponse"
                }
              }
            }
          },
          "404": {
            "description": "Saving bag not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/saving-bags/{bag_id}/answer": {
      "post": {
        "tags": [
          "saving-bags"
        ],
        "summary": "Submit answers to saving bag questions (REQ-API-05)",
        "description": "Submit user answers to the saving bag questionnaire (REQ-API-05).\n\n- Persists answers via the `savings` MCP server.\n- Researches real itemized costs (using Gemini grounding or static price table fallback).\n- Computes mathematical feasibility against user's actual cash flow (INV-015).\n- Emits a dated, funded plan in A2UI with milestone breakdown.",
        "operationId": "answer_saving_bag",
        "parameters": [
          {
            "name": "bag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingBagAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answers processed; research and deterministic feasibility plan generated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingBagResponse"
                }
              }
            }
          },
          "404": {
            "description": "Saving bag not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/saving-bags/{bag_id}/refresh": {
      "post": {
        "tags": [
          "saving-bags"
        ],
        "summary": "Refresh saving bag research (REQ-API-05)",
        "description": "Refresh price research and recalculate feasibility plan for a saving bag (REQ-API-05).",
        "operationId": "refresh_saving_bag",
        "parameters": [
          {
            "name": "bag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingBagRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Research refreshed and plan recomputed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingBagResponse"
                }
              }
            }
          },
          "404": {
            "description": "Saving bag not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status ('ok' or 'error').",
            "examples": [
              "ok"
            ]
          },
          "surface_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Id",
            "description": "Identifier of the active or newly persisted A2UI surface.",
            "examples": [
              "surf_1a2b3c4d"
            ]
          },
          "a2ui": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "A2Ui",
            "description": "Full A2UI message array for this turn (INV-010, INV-016)."
          },
          "assistant_text": {
            "type": "string",
            "title": "Assistant Text",
            "description": "Text utterance or commentary from the AI agent.",
            "default": "",
            "examples": [
              "He analizado tus deudas y reestructurado el plan."
            ]
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues",
            "description": "List of warnings, schema validation issues, or soft errors."
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Optional human-readable error or status message."
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor",
            "description": "Current acting party in negotiation ('user' or 'bank').",
            "examples": [
              "bank"
            ]
          },
          "round": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Round",
            "description": "Current negotiation round index.",
            "examples": [
              1
            ]
          },
          "catalog_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id",
            "description": "Catalog ID used to render the interface ('amitie.standard.v1' or 'amitie.voz_color.v1').",
            "examples": [
              "amitie.standard.v1"
            ]
          },
          "audio_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Ref",
            "description": "URL to stream generated TTS audio for voice accessibility (REQ-ACC-03).",
            "examples": [
              "/api/audio/asset_123"
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "Structured error code (e.g. 'provider_unavailable', 'bad_request')."
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "description": "Indicates whether the client may retry this turn (REQ-NFR-04).",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AgentResponse",
        "description": "Standard agent turn response carrying the generated A2UI interface and metadata.",
        "examples": [
          {
            "a2ui": [
              {
                "catalogId": "amitie.standard.v1",
                "root": "root",
                "surfaceId": "surf_1a2b3c4d",
                "type": "beginRendering"
              }
            ],
            "assistant_text": "He analizado tus deudas y reestructurado el plan.",
            "audio_ref": "/api/audio/asset_123",
            "catalog_id": "amitie.standard.v1",
            "issues": [],
            "retryable": false,
            "status": "ok",
            "surface_id": "surf_1a2b3c4d"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HttpErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation of why the request failed.",
            "examples": [
              "unknown session"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HttpErrorDetail",
        "description": "Standard HTTP error response payload.",
        "examples": [
          {
            "detail": "unknown session"
          }
        ]
      },
      "MessageRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Active conversation session ID.",
            "examples": [
              "sess_demo"
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "User message in natural language. Required if audio_b64 is omitted.",
            "examples": [
              "Tengo cinco deudas y no me alcanza el mes"
            ]
          },
          "audio_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio B64",
            "description": "Base64-encoded audio. Transcribed via voice MCP when text is absent (REQ-ACC-04)."
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "BCP-47 language tag for audio transcription.",
            "default": "es-MX",
            "examples": [
              "es-MX"
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "MessageRequest",
        "description": "Payload containing the user's natural language or voice utterance.",
        "examples": [
          {
            "language": "es-MX",
            "session_id": "sess_demo",
            "text": "Tengo cinco deudas y no me alcanza el mes"
          }
        ]
      },
...
      "SavingBagCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Owner user identifier.",
            "default": "u_ana",
            "examples": [
              "u_ana"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of the savings goal (e.g. 'Viaje a Japón').",
            "examples": [
              "Viaje a Japón"
            ]
          },
          "target_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount",
            "description": "User-estimated target cost in MXN.",
            "examples": [
              75000.0
            ]
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date",
            "description": "Target completion date in ISO format (YYYY-MM-DD).",
            "examples": [
              "2027-04-15"
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional active session to bind generated UI to.",
            "examples": [
              "sess_demo"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavingBagCreateRequest",
        "description": "Payload to initiate a new Saving Bag goal.",
        "examples": [
          {
            "name": "Viaje a Japón",
            "session_id": "sess_demo",
            "target_amount": 75000.0,
            "target_date": "2027-04-15",
            "user_id": "u_ana"
          }
        ]
      },
...
      "SavingBagResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status ('ok' or 'error').",
            "examples": [
              "ok"
            ]
          },
          "bag_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bag Id",
            "description": "Saving bag identifier.",
            "examples": [
              "bag_123"
            ]
          },
          "bag": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bag",
            "description": "Saving bag entity attributes."
          },
          "bags": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Bags",
            "description": "List of saving bags (populated when listing bags)."
          },
          "plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan",
            "description": "Calculated financial feasibility plan (INV-015)."
          },
          "surface_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Id",
            "description": "Active A2UI surface ID for the saving bag."
          },
          "a2ui": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "A2Ui",
            "description": "Full A2UI message array for rendering."
          },
          "assistant_text": {
            "type": "string",
            "title": "Assistant Text",
            "description": "Agent commentary and explanation.",
            "default": ""
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues",
            "description": "Issues or warnings encountered."
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable status or error message."
          },
          "catalog_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id",
            "description": "A2UI catalog identifier."
          },
          "audio_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Ref",
            "description": "Audio stream reference URL."
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "Structured error code."
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "description": "Whether this operation is retryable.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SavingBagResponse",
        "description": "Composite response for Saving Bags operations including metadata and A2UI plan.",
        "examples": [
          {
            "a2ui": [],
            "assistant_text": "He investigado los costos reales y calculado tu plan de ahorro.",
            "bag": {
              "id": "bag_123",
              "name": "Viaje a Japón",
              "target_amount": 75000.0
            },
            "bag_id": "bag_123",
            "bags": [],
            "issues": [],
            "plan": {
              "feasible": true,
              "monthly_saving": 4500.0,
              "months": 16
            },
            "status": "ok",
            "surface_id": "surf_bag123"
          }
        ]
      },
...
    }
  },
...
}